+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 - added IGW
 - one/multiples NAT GW(s)
 - route tables for publics subnet and private internet subnets
+## [v1.1.0] Added
+- added security group module
+- modified README.md for each module